+++ bus.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void init_memory(void);
-void load_memory(uint16_t addr, uint8_t *data, uint16_t size);
+void load_memory(uint16_t addr, uint8_t *data, uint32_t size);
 
 #endif
